moved test setup in different file
